made rule parsing more efficient
The old for-loop required you to iterate thru every category key, and on every single iteration, it rewrote inp, hoping that it found the correct category. This was a huge waste of time and probably memory too, from looping unnecessarily and also reassigning the inp variable on every single iteration. That's ESPECIALLY if inp doesn't even have categories, e.g. in the sound change '[sm]$>'

This commit changed the else statement, so that the loop doesn't occur if categories aren't even in inp. Since categories are marked by '%', we just check for that. 
If '%' is in inp, it iterates thru inp and knows when to replace, instead of iterating thru the entire categories dictionary and blindly rewriting inp on every single iteration.
So it iterates thru inp, checks for categories, and if it finds one, it replaces it with the appropriate regex range.
If '%' is NOT in inp, it just adds the tuple of (inp, out). No need to check/replace anything.

I did some basic timing tests with time.perf_counter(). I timed the for loop and added up each time it did the loop. My code bested at 0.000_073 sec, averages around 0.000_12-0.000_15 sec, and worsted at 0.000_236 sec. The old loop bested at 0.000_702 sec, averaged between 0.001-0.000_8 sec, and worsted at 0.001_9 sec. 

That's a nearly a whole magnitude of 10 faster on average. And the worst of the new is faster than the best of the old.

Even though this probably varies from machine to machine, there's no doubt this method is a lot faster and does less unnecessary checking/replacing. 

I tested with some sound changes, and checked the rules list (line 307) that the program makes, and they're all fine.